prodsync: don't produce XME from stale data
* Previously, modified records to sync were searched on the master DB,
but the XME generation was reading from the slave DB, leading to stale
information in case the slave had not been updated yet. This is fixed by
performing all searches on the slave. This can result in outdated
records being sent in the prodsync, but those will eventually be updated
in the next run of the prodsync.

Signed-off-by: Micha Moskovic <michamos@gmail.com>